Update instruments page styling
