+++ cypress/component/Instruments.cy.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import Instruments from '../../src/Instruments';
+import { mount } from '@cypress/react';
+
+describe('<Instruments />', () => {
+  beforeEach(() => {
+    mount(<Instruments />);
+  });
+});
