+++ src/HomePage.tsx
@@ -208,7 +208,7 @@ const HomePage = (): React.ReactElement => {
                 </PaperDescription>
                 <PaperDescription variant="body1">
                   <Trans i18nKey="home-page.browse.description2">
-                    <strong>Flexible Interactive Reduction</strong> focuses on providing scientists an interface to
+                    <strong>Flexible Interactive Automation</strong> focuses on providing scientists an interface to
                     perform automatic reductions for beamline instruments from the web.
                   </Trans>
                 </PaperDescription>
@@ -255,11 +255,9 @@ const HomePage = (): React.ReactElement => {
                 <Avatar sx={avatarStyles}>
                   <SearchIcon sx={avatarIconStyles} />
                 </Avatar>
-                <PaperHeading variant="h4">{t('Live reductions')}</PaperHeading>
+                <PaperHeading variant="h4">{t('ISIS instruments')}</PaperHeading>
                 <PaperDescription variant="body1">
-                  {t(
-                    'Browse the instrument page to select an instrument to access the live data functionality for a specific instrument.'
-                  )}
+                  {t("Browse a list of ISIS Neutron and Muon Source's instruments.")}
                 </PaperDescription>
                 <Box marginTop="auto">
                   <Button
@@ -281,9 +279,9 @@ const HomePage = (): React.ReactElement => {
                 <Avatar sx={avatarStyles}>
                   <SearchIcon sx={avatarIconStyles} />
                 </Avatar>
-                <PaperHeading variant="h4">{t('Historical reductions')}</PaperHeading>
+                <PaperHeading variant="h4">{t('Historic reductions')}</PaperHeading>
                 <PaperDescription variant="body1">
-                  {t('Browse and search a list of all reduction performed on the platform.')}
+                  {t('Browse a list of interactable reductions performed on this platform.')}
                 </PaperDescription>
                 <Box marginTop="auto">
                   <Button
